chore(atomic): move RedirectionPayload to common
https://coveord.atlassian.net/browse/KIT-4178
