+++ packages/atomic/src/components/common/search-box/redirection-payload.ts
